requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.4
django-extensions==3.2.1
django-on-heroku==1.1.2
django-storages==1.13.2
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.3.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
whitenoise==6.2.0
